Just A Prototype
Currently, the game has a market which fluctuates up and down, and
creates trends which determine the houses that show up. Houses are
formed every 10 days, and the trends on the market reset every month.
Currently the game is entirely text based. If the game continues to
receive work, good graphics are necessary.